requirements.txt
streamlit>=1.20
spacy>=3.5
pymupdf>=1.22.0        # pip package name: PyMuPDF
python-docx>=0.8.11
pandas>=2.0
regex
tabulate
